pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viper"
version = "0.1.4"
description = "VIPER: Viral Interpretation and Protein Evolution Reporter — mutation hotspot detection and functional impact annotation for viral genomes"
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "Andrew Vazzano"}]
dependencies = [
  "biopython>=1.79",
  "vcfpy>=0.14",
  "pandas>=1.5",
  "numpy>=1.24",
  "scipy>=1.9",
  "matplotlib>=3.6",
  "seaborn>=0.12"
]

[tool.setuptools.packages.find]
where = ["."]